Add the debug_test method to the scheduler
When running tests, we will always end up using stdout and stderr files
that are under control of the executor.  However, for tests executed in
debug mode, the output will probably go to the console -- and we need
some special code in that case (e.g. to be able to parse the stdout to
compute the test result).

Separate the entry points to regular test runs and debug runs so that we
can implement proper handling of output to the screen.