Modify handling of output files in debug_test
When running a test via debug_test, we cannot simply send its output to the
real stdout or stderr: the test interface's compute_result sometimes needs
access (see TAP) to the test's output in order to compute the result.  For
this to be possible, we need to send the output to our own files so they
can be reread for result computation, and separately we need to send the
files to wherever the user specified.

Because of this, change debug_test to use the executor-generated output files
internally and later copy the output to the user-provided output files.

This is still suboptimal because we now lose visibility of the output data
as it is generated... but I'll address this separately.