Revert a68577bbff7b9b21da2a2cdcb7d0930582dedf93
This change reverts the huge hack introduced in "Make sure TAP test
programs can be debugged" and fixes a regression with the handling
of stdout vs. stderr in TAP tests.  The scheduler interface has been
modified to carry a debug_test() method that exposes the proper interface
to support the behavior of TAP test programs, and we can now rely on
that instead.