Do not write out invalid characters to XML file
For special characters such as: <, >, or ', write them out to the file
as: &#[decimal ASCII value];

Character     Written to    Appears in Jenkins
to encode     XML file      test result viewer
=========     ==========    ==================
  <            &lt;             <
  >            &gt;             >
  '            &apos;           '
  &            &amp;            &
  "            &quot;           "

For characters which are listed as RestrictedChar in
http://www.w3.org/TR/xml11/#charsets , these characters are completely
invalid XML, and cannot even be escaped.

Character     Written to    Appears in Jenkins
to encode     XML file      test result viewer
=========     ==========    ==================
  0x08         &amp;#8;          &#8;
  0x1F         &amp;#31;         &#31;

This will at least generate a valid XML file, but let us see where these
restricted characters would appear in the output.

Fixes #136