Fix test when GDB is missing
The dump_stacktrace__cannot_find_core test, which attempts to cause a failure
due to the lack of a core file, would fail too early if there was no GDB binary
available (as is the case in recent OS X versions).

Fix this by creating a fake GDB so that we never trip over this condition
accidentally.