Preserve PATH when invoking sudo
The new Travis CI Precise infrastructure seems to enforce env_reset in the
sudo configuration and does not allow using the -E flag to bypass it.
Instead, explicitly pass the PATH to the 'make distcheck' sudo invocation
so that it can find the right tools.

This fixes the AS_ROOT=yes builds after we were silently switched to the
new infrastructure.