Allow getcwd calls in unprivileged tests
For getcwd to work, all path components leading to the current directory
must have read AND search permissions.  Up until now, we were only granting
search permissions and thus getcwd was broken for any test running as an
unprivileged user when Kyua was started as root.

As part of this change, rename all of our mkdtemp functions to mkdtemp_public
to clearly denote that they do something different to the standard mkdtemp.

This effectively fixes execution of unprivileged tests when running as root
in at least NetBSD 7.0 because our own code issues a getcwd() call as part
of the test case setup.  This may have been a regression in Kyua 0.12
because maybe we were not calling getcwd() during test setup in the past
(haven't checked), but what this means is that any test programs calling
getcwd might have mysteriously failed before.