Be more lax when processing TAP directives
According to http://testanything.org/tap-specification.html, it seems that
the SKIP and TODO directives need to be matched without respecting case
(yet the mention for this is buried in the text in a place you would not
expect, yay).  Make it so.

Additionally, one of the examples in that page shows the directive to skip
a test as "Skipped:", so make sure to also accept this and the variant
without the trailing colon... just in case.

Fixes #159.