Fix crash when the PID of a cleanup part is reused
Make sure we clean up the tracking data of subprocesses that correspond
to a test case's cleanup routine.  Otherwise we crash when the PID of
such objects is reused (only when it's reused for a non-cleaup routine,
but that's the most likely case) because we return an invalid PID to the
run_tests driver, and that one gets confused because it cannot find the
process in its own tracking table.

Not to mention that wait_any was returning bogus data in the
non-crashing cases to the caller; i.e. for the regular reuse of a PID
matching a test subprocess, it was returning data for a previous
execution.

While doing this, add more logging statements and invariants to catch
similar potential problems.  In particular, ensure we do not issue an
insert on an std::map that already contains the key we try to put
(because the "helpful" behavior of insert it so silently drop the given
value in this case, which is what exposed the crash here).

Fixes #156.