Avoid using references to hold pointers
Using a reference to keep a smart pointer around is bad because doing so
prevents the reference-counting mechanism from working.  This should not
be an issue in the remaining cases fixed here, but a different instance
fixed in the parent commit caused issues with the previous fix itself.
Better fix them all for consistency.