Fix possible crash when unprogramming timers
The signals::timer::unprogram method asserted that the timer was still
programmed.  However, this condition may not hold true when the function
is entered if the timer expires at around the same time.

Observed an occasional crash in this code path but was not able to
reproduce again.  Make this more robust, which will hopefully prevent
this race condition.