Fix override_all_with_set_string on 32-bit
Explicitly cast a large literal integer to 64-bit so that it can properly
represent a bytes quantity.  This was the intent in the previous code due
to the presence of the 'L' qualifier... but this was incorrect because
long does not imply 64-bit on 32-bit machines.  Therefore, just cast to
uint64_t.

This error was actually triggering a warning on 32-bit builds but was not
caught because we do not run 32-bit builds on Travis CI.

Fixes #134.