Enable C++11 support
To do so, enable support for the C++11 standard during configure and
sanity-check that it works.

Then, and as a first step, change the code to drop the shared_ptr.hpp
compatibility header as proof that this is working.  I'm doing this to
allow the codebase of using "new" (6-year old, at this point) features.
In particular, I want to be able to use threads to fix the parallel
execution of test cases.

Fixes #152.