Redo timeout enforcement using a thread
Replace all the extremely complex logic involved in programming system
timers to enforce test timeouts.  Instead, just create a thread that
wakes once a second and kills any tests whose timeout has expired.
The granularity here is sufficient and the code is much simpler.

The reason for this change is to stop having to handle SIGALRM at all,
which makes the migration to a threaded executor more difficult.  We
could use timers again in the future if the constant waking of the
thread turns out to be a problem, but we should only do that once we
have sorted out signal handling with threads first.