Redo interrupt handling using a thread
Funnel all signal handling through a single thread.  This is a necessary
preparation step to permit spawning more threads in Kyua because we need
to properly control how signals are handled.

This change makes it possible to properly detect double signal delivery
and to terminate abruptly in that case.  Previous versions of Kyua would
not handle this case properly, possibly entering a state where they
could not be gracefully killed.