Merge test_result_handle into result_handle
Originally, I think I separated the two concepts of a generic process
termination handle from the test results because I was planning on making
the list_test_cases function return a result_handle as well.  However,
this won't really be the case any longer with a threaded model, so drop
the split for simplicity.