Fix running `executor_pid_test` on ZFS
The testcases prior to this change relied on setting UF_NOUNLINK, which
is unfortunately bugged on ZFS: ZFS does not treat UF_NOUNLINK as a valid
value for fflags, which causes the testcases to fail.

Switch to SF_NOUNLINK. This particular fflag works work ZFS, unlike
UF_NOUNLINK, but requires root privileges, so add the needed metadata
to the Kyuafile.

As part of this, I needed to convert the test from atf-c++(3) to
atf-c(3). The atf-c++(3) APIs do not have the printf-like equivalent
formatting, so it was very difficult figuring out why things weren't
working in atf-c++(3).

While here, clean up the code to DRY some of the repeated logic that it
contained.

The UF_NOUNLINK bug has been reported in https://github.com/openzfs/zfs/issues/16809 .

Fixes:  e26c8047f7299f046f87308705b062973d686f5f