Add more information to failed results
This change adds more information to "premature failures" or failures
where no output was provided on the command-line. This is being done to
help make the reasons for why a failure occurred clearer to the
end-user.

This also fixes the fact that several ATF-related tests were failing with
"empty output" (lacking exit code data), or not failing with the
appropriate (expected) "timeout" messages.

While here, address the fact that newer versions of sqlite3 fail when
executing statements with `sqlite::api_error` instead of `store::error`.

This addresses part of #237 and makes the kyua tests pass cleanly
again on FreeBSD.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>