Fix circular dependency post-execenv introduction
Prior to this change, when this code would compile it would result in
linking failures when using binutils' ld on Ubuntu 24.04. This issue
did not reproduce outside of this scenario.

Instead of model/... relying on the execenv objects, take the single
dependency, turn it into a proper constant and export it via
`engine/execenv/execenv_fwd.hpp`.

This unbreaks the remaining build issue on Ubuntu 24.04.

Fixes:	a4737b147c3111b960531db4dcf724936e4f81a5

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>