integration tests: don't forkbomb in my MacOS environment
Prior to this change the tests in my environment (where kyua isn't in
my default `$PATH`) would forkbomb my MacOS environment, resulting in
testcase failures and other resource problems.

Capture the path to the actual chosen kyua binary and use that in the
dummy kyua wrapper scripts which massage the output to be deterministic.
This avoids problems where the dummy scripts would call themselves in a
recursive manner (previously).

Closes:	#236

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>