integration tests: don't forkbomb in my MacOS environment
Prior to this change the tests in my environment would forkbomb my MacOS environment,
resulting in testcase failures and other system resource problems. This was happening,
because of how the integration tests hijack the kyua command and inject the dummy kyua
commands into the `$PATH`, resulting in the dummy kyua scripts calling themselves
recursively.

Capture the path to the actual chosen kyua binary and use that in the dummy kyua wrapper
scripts which massage the output to be deterministic instead of relying on the first
entry in $PATH not being the dummy kyua wrapper scripts.

Closes:	#236

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>